move seo into a plugin
resolves #90
