+++ plugins/gatsby-ample-seo/README.md
@@ -1,17 +1,5 @@
-import { Meta, Preview, Props, Story } from "@storybook/addon-docs/blocks"
-
-import SEO from "./"
-
-<Meta title="Utilities/SEO" component={SEO} />
-
 # SEO
 
 The SEO component is used by template adapters to render meta tags into the head of the page.
 
 This utility takes normalized SEO data from the page level, then interjects global defaults, such as a title template and a fallback image. It then does a bit more cleanup and passes the data onto the SEO component for it to render the meta tags in the head of the page. Note that the global defaults are configurable in the CMS under _Config_ > _SEO_.
-
-## Properties
-
-The props for the local utility are listed below. But, as long as you're using the `normalizeSEO` helper in the adapter, these you won't have to worry about structuring everything perfectly for this component.
-
-<Props of={SEO} />
