+++ plugins/gatsby-ample-seo/index.js
@@ -1,0 +1,4 @@
+import normalizeSEO from "./src/helpers/normalize-seo"
+import SEO from "./src/components/seo"
+
+export { normalizeSEO, SEO }
