+++ plugins/gatsby-ample-seo/package.json
@@ -1,0 +1,1 @@
+{}
