+++ plugins/gatsby-ample-seo/src/helpers/normalize-seo/index.js
