--- src/helpers/index.js
@@ -1,3 +1,0 @@
-import normalizeSEO from "./normalize-seo"
-
-export { normalizeSEO }
